3.0.0 extreme tests (#839)
* move rtoken extreme tests to their own file and fix.

* fix run script.

* fix cases.

* fix linting.

* add expectEvents.

* fix test:extreme command.
