+++ scripts/exhaustive-tests/run-1.sh
@@ -1,3 +1,3 @@
 echo "Running RToken & Furnace exhaustive tests for commit hash: "
 git rev-parse HEAD;
-NODE_OPTIONS=--max-old-space-size=30000 EXTREME=1 SLOW=1 PROTO_IMPL=1 npx hardhat test test/RToken.test.ts test/Furnace.test.ts;
+NODE_OPTIONS=--max-old-space-size=30000 EXTREME=1 SLOW=1 PROTO_IMPL=1 npx hardhat test test/RTokenExtremes.test.ts test/Furnace.test.ts;
